[rrc] split the front leg into driver::frontend
frontend(), frontend_package(), the textual-dump re-entry helpers
(refetch_sources), and the shared finish (lower_or_render, finish_mir)
move together. driver.rs is now the dispatch spine: main(), run(), the
package-root plumbing, and scan-deps.
